refactor(cake_tinygemm2): flatten csrc file, fold jit gen into tinygemm2.py, reuse is_sm100a_supported
- csrc/tinygemm2_sm100/tinygemm2_sm100.cu -> csrc/tinygemm2_sm100.cu: after
  the single-TU merge the subdirectory held one file; the incumbent
  convention is a flat csrc/tinygemm2.cu. Pure rename, content unchanged.
- gen_tinygemm2_sm100_module moved into flashinfer/jit/tinygemm2.py next to
  gen_tinygemm2_module (family jit files hold all their variants — see
  jit/fused_moe.py with seven gens); flashinfer/jit/tinygemm2_sm100.py
  removed, package-level import path unchanged.
- _use_tinygemm2_sm100 now gates on flashinfer.utils.is_sm100a_supported
  instead of a hand-rolled compute-capability tuple check — same devices,
  plus the util's CUDA>=12.8 toolchain floor the ad-hoc check was missing.
